Some reworking of the lexer. In order to deal with tokens spanning multiple lines `fgets` is clearly inappropriate (although it could be made to work, with multiple calls).
Instead, since this application must read the entire input anyway, I've decided to lex the entire input from one giant string. This breaks my heart, because it feels so excessive. However it does have several benefits (notably much error handling may be made redundant in a future commit).

In order to implement this, I'm using `mmap()` to get the file contents. This is just for convenience; it could equally well be a read loop.